split support for aeson into a separate package `quantification-aeson`
also narrowed dependency and extension lists
