+++ quantification-aeson/quantification-aeson.cabal
@@ -1,12 +1,12 @@
 cabal-version: 2.4
-name: quantification
+name: quantification-aeson
 version: 0.8
-synopsis: Rage against the quantification
+synopsis: Rage against the quantification - integration with Aeson
 description: Data types and typeclasses to deal with universally and existentially quantified types
 homepage: https://github.com/andrewthad/quantification#readme
 license: BSD-3-Clause
 license-file: LICENSE
-author: Andrew Martin
+author: Andrew Martin, Marseille Bouchard
 maintainer: andrew.thaddeus@gmail.com
 copyright: 2018 Andrew Martin
 category: Web
@@ -15,17 +15,14 @@ build-type: Simple
 library
   hs-source-dirs: src
   exposed-modules:
-    Data.Binary.Lifted
-    Data.Exists
-    Data.Monoid.Lifted
+    Data.Exists.Aeson
   build-depends:
+    , aeson >= 2.2 && < 2.3
     , base >= 4.11.1 && < 5
-    , binary >= 0.8 && < 0.11
-    , hashable >= 1.4 && < 1.5
-    , text >= 1.0 && < 2.2
-    , path-pieces >= 0.2 && < 0.3
     , containers >= 0.5 && < 0.7
+    , quantification >=0.8 && <0.9
     , unordered-containers >= 0.2 && < 0.3
+    , vector >= 0.11 && < 0.14
   default-language: Haskell2010
   ghc-options: -O2 -Wall
